Add RowId as key column to table and move display column to ExpData table
